test(#zimic): unify dynamic path tests by method
