+++ packages/zimic/src/interceptor/http/interceptor/__tests__/shared/methods/head.ts
@@ -420,97 +420,6 @@ export function declareHeadHttpInterceptorTests(options: RuntimeSharedHttpInterc
     });
   });
 
-  describe('Dynamic paths', () => {
-    it('should support intercepting HEAD requests with a dynamic path', async () => {
-      await usingHttpInterceptor<{
-        '/users/:id': {
-          HEAD: {
-            response: {
-              200: {};
-            };
-          };
-        };
-      }>(interceptorOptions, async (interceptor) => {
-        const genericHeadHandler = await promiseIfRemote(
-          interceptor.head('/users/:id').respond((request) => {
-            expectTypeOf(request.pathParams).toEqualTypeOf<{ id: string }>();
-            expect(request.pathParams).toEqual({ id: '1' });
-
-            return {
-              status: 200,
-            };
-          }),
-          interceptor,
-        );
-        expect(genericHeadHandler).toBeInstanceOf(Handler);
-
-        let genericHeadRequests = await promiseIfRemote(genericHeadHandler.requests(), interceptor);
-        expect(genericHeadRequests).toHaveLength(0);
-
-        const genericHeadResponse = await fetch(joinURL(baseURL, `/users/${1}`), { method: 'HEAD' });
-        expect(genericHeadResponse.status).toBe(200);
-
-        genericHeadRequests = await promiseIfRemote(genericHeadHandler.requests(), interceptor);
-        expect(genericHeadRequests).toHaveLength(1);
-        const [genericHeadRequest] = genericHeadRequests;
-        expect(genericHeadRequest).toBeInstanceOf(Request);
-
-        expectTypeOf(genericHeadRequest.pathParams).toEqualTypeOf<{ id: string }>();
-        expect(genericHeadRequest.pathParams).toEqual({ id: '1' });
-
-        expectTypeOf(genericHeadRequest.body).toEqualTypeOf<null>();
-        expect(genericHeadRequest.body).toBe(null);
-
-        expectTypeOf(genericHeadRequest.response.status).toEqualTypeOf<200>();
-        expect(genericHeadRequest.response.status).toEqual(200);
-
-        expectTypeOf(genericHeadRequest.response.body).toEqualTypeOf<null>();
-        expect(genericHeadRequest.response.body).toBe(null);
-
-        await promiseIfRemote(genericHeadHandler.bypass(), interceptor);
-
-        const specificHeadHandler = await promiseIfRemote(
-          interceptor.head(`/users/${1}`).respond((request) => {
-            expectTypeOf(request.pathParams).toEqualTypeOf<{ id: string }>();
-            expect(request.pathParams).toEqual({});
-
-            return {
-              status: 200,
-            };
-          }),
-          interceptor,
-        );
-        expect(specificHeadHandler).toBeInstanceOf(Handler);
-
-        let specificHeadRequests = await promiseIfRemote(specificHeadHandler.requests(), interceptor);
-        expect(specificHeadRequests).toHaveLength(0);
-
-        const specificHeadResponse = await fetch(joinURL(baseURL, `/users/${1}`), { method: 'HEAD' });
-        expect(specificHeadResponse.status).toBe(200);
-
-        specificHeadRequests = await promiseIfRemote(specificHeadHandler.requests(), interceptor);
-        expect(specificHeadRequests).toHaveLength(1);
-        const [specificHeadRequest] = specificHeadRequests;
-        expect(specificHeadRequest).toBeInstanceOf(Request);
-
-        expectTypeOf(specificHeadRequest.pathParams).toEqualTypeOf<{ id: string }>();
-        expect(specificHeadRequest.pathParams).toEqual({});
-
-        expectTypeOf(specificHeadRequest.body).toEqualTypeOf<null>();
-        expect(specificHeadRequest.body).toBe(null);
-
-        expectTypeOf(specificHeadRequest.response.status).toEqualTypeOf<200>();
-        expect(specificHeadRequest.response.status).toEqual(200);
-
-        expectTypeOf(specificHeadRequest.response.body).toEqualTypeOf<null>();
-        expect(specificHeadRequest.response.body).toBe(null);
-
-        const unmatchedHeadPromise = fetch(joinURL(baseURL, `/users/${2}`), { method: 'HEAD' });
-        await expectFetchError(unmatchedHeadPromise);
-      });
-    });
-  });
-
   describe('Restrictions', () => {
     it('should support intercepting HEAD requests having headers restrictions', async () => {
       type UserHeadHeaders = HttpSchema.Headers<{
